# 02-development/homework/balancio/openapi.yaml
openapi: 3.0.3
info:
  title: Balancio API
  description: >
    Shared expense splitter for friends, households, couples, trips, and
    other small groups. See _docs/specs.md for the full product spec.
  version: "0.1.0"
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
  /auth/setup:
    post:
      tags: [auth]
      summary: Create the first global admin account
      description: Only available while no admin exists yet.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupRequest"
      responses:
        "201":
          description: Admin created and logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Setup has already been completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/login:
    post:
      tags: [auth]
      summary: Log in with username and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Incorrect username or password, or account deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/change-password:
    post:
      tags: [auth]
      summary: Change the current user's password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /auth/me:
    get:
      tags: [auth]
      summary: Get the current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /categories:
    get:
      tags: [categories]
      summary: List categories (global, plus a group's own if group_id given)
      parameters:
        - name: group_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      tags: [categories]
      summary: Create a category (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /categories/{category_id}:
    parameters:
      - $ref: "#/components/parameters/CategoryId"
    patch:
      tags: [categories]
      summary: Rename a category (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [categories]
      summary: Delete a category (admin only, only when unused)
      responses:
        "204":
          description: Category deleted
        "400":
          description: Category is used by existing expenses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /{transaction_type}/{transaction_id}/comments:
    parameters:
      - name: transaction_type
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/TransactionType"
      - name: transaction_id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [comments]
      summary: List comments on a transaction
      responses:
        "200":
          description: List of comments, oldest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [comments]
      summary: Add a comment (only users involved in the transaction)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body:
                  type: string
                  minLength: 1
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /comments/{comment_id}:
    parameters:
      - $ref: "#/components/parameters/CommentId"
    patch:
      tags: [comments]
      summary: Edit a comment (author only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body:
                  type: string
                  minLength: 1
      responses:
        "200":
          description: Updated comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [comments]
      summary: Delete a comment permanently (author only)
      responses:
        "204":
          description: Comment deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    CategoryId:
      name: category_id
      in: path
      required: true
      schema:
        type: string
...
    CommentId:
      name: comment_id
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Validation error or invalid financial state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated but not allowed to perform this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Friendly, domain-specific message. Never a stack trace.
    Role:
      type: string
      enum: [admin, user]
    Theme:
      type: string
      enum: [light, dark]
...
    User:
      type: object
      required:
        - id
        - username
        - display_name
        - role
        - is_active
        - must_change_password
        - theme
        - created_at
        - updated_at
      properties:
        id:
          type: string
        username:
          type: string
        display_name:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        is_active:
          type: boolean
        must_change_password:
          type: boolean
        theme:
          $ref: "#/components/schemas/Theme"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SetupRequest:
      type: object
      required: [username, display_name, password]
      properties:
        username:
          type: string
        display_name:
          type: string
        password:
          type: string
          minLength: 8
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, access_token, token_type]
      properties:
        user:
          $ref: "#/components/schemas/User"
        access_token:
          type: string
        token_type:
          type: string
          enum: [bearer]
    ChangePasswordRequest:
      type: object
      required: [old_password, new_password]
      properties:
        old_password:
          type: string
        new_password:
          type: string
          minLength: 8
...
    Category:
      type: object
      required: [id, name, group_id, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        group_id:
          type: string
          nullable: true
          description: null means a global category.
        created_at:
          type: string
          format: date-time
    CreateCategoryRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        group_id:
          type: string
          nullable: true
...
    TransactionType:
      type: string
      enum: [expense, settlement, refund]
    Comment:
      type: object
      required:
        - id
        - transaction_type
        - transaction_id
        - author_id
        - body
        - created_at
        - updated_at
      properties:
        id:
          type: string
        transaction_type:
          $ref: "#/components/schemas/TransactionType"
        transaction_id:
          type: string
        author_id:
          type: string
        body:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...